1.3.0 style: ライブラリ名を変更
