--- Tools/control/Pid/Pid.hpp
@@ -1,13 +1,0 @@
-#ifndef _PID_HPP_
-#define _PID_HPP_
-#include <mbed.h>
-
-class Pid
-{
-    public:
-    Pid(double, double, double);
-    double calculate(double, double);
-    double Kp, Ki, Kd;
-};
-
-#endif
